+++ src/main/java/com/sparta/trello/common/exception/ExceptionDto.java
@@ -1,4 +1,4 @@
-package com.sparta.trello.exception;
+package com.sparta.trello.common.exception;
 
 import lombok.Builder;
 import lombok.Getter;
