+++ src/main/java/com/sparta/trello/common/jwt/JwtAuthorizationFilter.java
@@ -1,7 +1,6 @@
-package com.sparta.trello.jwt;
+package com.sparta.trello.common.jwt;
 
 import com.sparta.trello.auth.entity.Role;
-import com.sparta.trello.auth.security.UserDetailsImpl;
 import com.sparta.trello.auth.security.UserDetailsServiceImpl;
 import io.jsonwebtoken.Claims;
 import jakarta.servlet.FilterChain;
